+++ backend/canisters/user/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Remove a few deprecated methods ([#4006](https://github.com/open-chat-labs/open-chat/pull/4006))
+- Consolidate remove and block community permissions ([#4030](https://github.com/open-chat-labs/open-chat/pull/4030))
 
 ### Fixed